# collab-service/internal/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.FolderDTO:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  models.Folder:
    properties:
      id:
        type: string
      name:
        type: string
      notes:
        items:
          $ref: '#/definitions/models.Note'
        type: array
      owner:
        $ref: '#/definitions/models.User'
      ownerID:
        type: string
    type: object
  models.Note:
    properties:
      body:
        type: string
      folder:
        $ref: '#/definitions/models.Folder'
      folderID:
        type: string
      id:
        type: string
      owner:
        $ref: '#/definitions/models.User'
      ownerID:
        type: string
      title:
        type: string
    type: object
  models.Roster:
    properties:
      id:
        type: string
      isLeader:
        type: boolean
      team:
        $ref: '#/definitions/models.Team'
      teamID:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: string
    type: object
  models.Team:
    properties:
      id:
        type: string
      rosters:
        items:
          $ref: '#/definitions/models.Roster'
        type: array
      teamName:
        type: string
    type: object
  models.User:
    properties:
      email:
        type: string
      folders:
        items:
          $ref: '#/definitions/models.Folder'
        type: array
      id:
        type: string
      passwordHash:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      rosters:
        items:
          $ref: '#/definitions/models.Roster'
        type: array
      username:
        type: string
    type: object
  models.UserRole:
    enum:
    - MANAGER
    - MEMBER
    type: string
    x-enum-varnames:
    - RoleManager
    - RoleMember
host: localhost:8080
info:
  contact: {}
  description: API for managing teams, folders, notes, and sharing.
  title: Collab Service API
  version: "1.0"
paths:
  /folders:
    post:
      consumes:
      - application/json
      description: Create a new folder for the authenticated user
      parameters:
      - description: Folder data
        in: body
        name: folder
        required: true
        schema:
          $ref: '#/definitions/dto.FolderDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created folder
          schema:
            $ref: '#/definitions/models.Folder'
        "400":
          description: Bad request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: User not found
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create a new folder
      tags:
      - folders
  /folders/{folderId}:
    delete:
      description: Delete a folder and all its notes
      parameters:
      - description: Folder ID
        in: path
        name: folderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Folder not found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Delete folder
      tags:
      - folders
    get:
      description: Retrieve details of a specific folder
      parameters:
      - description: Folder ID
        in: path
        name: folderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Folder details
          schema:
            $ref: '#/definitions/models.Folder'
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Folder not found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get folder details
      tags:
      - folders
    put:
      consumes:
      - application/json
      description: Update a folder's name
      parameters:
      - description: Folder ID
        in: path
        name: folderId
        required: true
        type: string
      - description: Updated folder data
        in: body
        name: folder
        required: true
        schema:
          $ref: '#/definitions/dto.FolderDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Updated folder
          schema:
            $ref: '#/definitions/models.Folder'
        "400":
          description: Bad request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Folder not found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update folder
      tags:
      - folders
...
  /teams/{teamId}/assets:
    get:
      description: 'Manager-only API: View all assets that team members own or can
        access (shared folders or notes)'
      parameters:
      - description: Team ID
        in: path
        name: teamId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response with owned and shared folders
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error response
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all accessible assets of team members
      tags:
      - assets
...
  /users/{userId}/assets:
    get:
      description: Get all folders owned or shared with a specific user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response with owned and shared folders
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error response
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all assets of a user
      tags:
      - assets
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT token like: Bearer <your-token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"